FIX sglang can only operate in float16 mode. Allow bfloat16

+++ python/sglang/srt/layers/token_attention.py
@@ -11,8 +11,9 @@ if global_server_args_dict.get("attention_reduce_in_fp32", False):
     REDUCE_TRITON_TYPE = tl.float32
     REDUCE_TORCH_TYPE = torch.float32
 else:
-    REDUCE_TRITON_TYPE = tl.float16
-    REDUCE_TORCH_TYPE = torch.float16
+    # TODO FIX ME. this should match model ServerArgs dtype
+    REDUCE_TRITON_TYPE = tl.bfloat16
+    REDUCE_TORCH_TYPE = torch.bfloat16
 
 
 @triton.jit
